+++ rts/motoko-rts-tests/src/gc/classical.rs
@@ -64,13 +64,11 @@ impl GC {
             }
 
             GC::Generational => {
-                use motoko_rts::gc::{
-                    generational::{
-                        write_barrier::{LAST_HP, REMEMBERED_SET},
-                        GenerationalGC, Strategy,
-                    },
-                    remembered_set::RememberedSet,
+                use motoko_rts::gc::generational::{
+                    write_barrier::{LAST_HP, REMEMBERED_SET},
+                    GenerationalGC, Strategy,
                 };
+                use motoko_rts::gc::remembered_set::RememberedSet;
 
                 let strategy = match _round {
                     0 => Strategy::Young,