+++ rts/motoko-rts-tests/src/gc/incremental/roots.rs
@@ -136,6 +136,9 @@ unsafe fn get_roots(heap: &MotokoHeap) -> Roots {
         unused_root,
         unused_root,
         unused_root,
+        unused_root,
+        unused_root,
+        unused_root,
     ]
 }